React: moved not_found into it's own component and fixed bug in profile
