+++ client/src/components/not_found/index.jsx
@@ -1,14 +1,19 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import styled from 'styled-components';
 
-import './not-found.scss';
+const Container = styled.div`
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+`;
 
 const NotFound = () => (
-  <div className="NotFound">
+  <Container>
     <img src="/images/404.png" />
     <h2>Not Found</h2>
     <Link to="/">Get me outta here!</Link>
-  </div>
+  </Container>
 );
 
 export default NotFound;
