--- client/src/navigation/not-found.scss
@@ -1,5 +1,0 @@
-.NotFound {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-}
